feat: add admission request detail page
